Subjacobian working well. Moved I/O stuff into new file
