+++ src/AGNI.jl
@@ -19,6 +19,7 @@ module AGNI
     include("spectrum.jl")
     include("atmosphere.jl")
     include("setpt.jl")
+    include("saveload.jl")
     include("plotting.jl")
     include("energy.jl")
     include("solver.jl")
@@ -27,6 +28,7 @@ module AGNI
     import .phys
     import .atmosphere
     import .setpt
+    import .saveload
     import .plotting 
     import .energy
     import .solver 
@@ -362,7 +364,7 @@ module AGNI
         @info prt_req[1:end-2]
 
         # Write initial state
-        atmosphere.write_pt(atmos, joinpath(atmos.OUT_DIR,"pt_ini.csv"))
+        saveload.write_pt(atmos, joinpath(atmos.OUT_DIR,"pt_ini.csv"))
 
         # Do chemistry on initial composition
         if chem_type in [1,2,3]
@@ -428,9 +430,9 @@ module AGNI
 
         # Write arrays
         @info "Writing results"
-        atmosphere.write_pt(atmos,      joinpath(atmos.OUT_DIR,"pt.csv"))
-        atmosphere.write_ncdf(atmos,    joinpath(atmos.OUT_DIR,"atm.nc"))
-        atmosphere.write_fluxes(atmos,  joinpath(atmos.OUT_DIR,"fl.csv"))
+        saveload.write_pt(atmos,      joinpath(atmos.OUT_DIR,"pt.csv"))
+        saveload.write_ncdf(atmos,    joinpath(atmos.OUT_DIR,"atm.nc"))
+        saveload.write_fluxes(atmos,  joinpath(atmos.OUT_DIR,"fl.csv"))
 
         # Save plots
         @info "Plotting results"
